requirements.txt
# dev
isort
black
pylama
pydocstyle
pytest-cov
mccabe
flaky

# prod
boto3
transformers
huggingface_hub
torch
